changes_2013_03_18c.patch
This set of patches does the following:

1.  Fixed a typo ( where "+ sizeof()"  should have been "* sizeof()")
which caused a memory problem for EMF/WMF files with very large numbers
of hatches.

2. Added support for background mode, background color, and textcolor
in hatches.  EMF/WMF files change these parameters but the change may
be silent until many records later.  This has the odd effect that a
stroke or fill may be defined (in SVG) and then it is ignored later
and replaced with one with a different background color. 

3. Fixed WMF output so that it wasn't adding +1 to the number of pixels
for Width and Height.  (Allows WMF files to go through several cycles
of save as, open without changing sizes.)

4.  Cleaned up indenting of [ew]mf-{print|inout}.* files, to make
them compliant with the inkscape standard.  All indents are (I hope)
now 4*i deep.

5.  Added underline/strikeout support for text read from EMF/WMF files.
Inkscape itself cannot currently render this, but it makes it into the
SVG, and it shows up correctly when that SVG is read by Opera.

6.  Took out all the "throw" calls, replacing them with g_error().  If
this comes up again in debugging a define can be used to remap the
g_error to temporarily reintroduce the throw so that gdb can catch them.

7.  Took out the "%6lf" format changes from patch 11724, retained the
"127" length limit in the same sscanf.

8.  Put the C type casts back in, reversing those changes from patch
11724.   This is a style issue, and I could not find clear guidance for
which way to go.  (Nor a good rationale for keeping the lengthier C++
syntax.)  So  I reviewed a large swath of other inkscape code to see if
there was a trend and found a very large number of other sections that
were using C style casts instead of the more verbose C++ forms.   So I
kept it the way it has been.

9.  The locale changes from 11724 were of course retained.

(bzr r11668.1.59)