Add command line option to specify PDF export version. Patch from schwieni.
From the patch's author:

Additional command line option added to export to different PDF-versions
supported by cairo (currently only "PDF 1.4" and "PDF 1.5"). The user must
provide the exact string found in the PDF-export dialog of Inkscape. This
feature was previously only accessible via the GUI of Inkscape. This option is
useful for people opting for PDF-a conformance in their PDFs. Moreover, the
Extension and Parameter classes are extended by setters for enum parameters
(used in combo boxes).

(bzr r12348)