btool:  Fix hitting cmdline limit on XP/Mingw/Msys (Fixes #1251405)
Current releases of trunk fail to build on Windows XP because command
line length limits are hit.  This patch sends only the unique parameters
from the "libs" field during the link stage to the command.  This saves
enough characters so that btool will complete the build.

It is not a long term solution, but will allow development to continue
on Windows for a while longer, until the underlying problems can be
addressed.

Patch authored by David Mathog (mathog)

Signed-off-by: Bryce Harrington <bryce@bryceharrington.org>

(bzr r12913)