Adding default argument to PathManipulator::isBSpline.
Default argument is used when function is called with missing arguments.
The default was chosen to 'true', because it is the safest value in this
context.

From now on, the function can be called like "_pm.isBSpline()", which is
more intuitive.

Still, I haven't studied the code enough to determine if this cache
behaviour (force the user to think if cache is correct) is appropriate.

(bzr r11950.7.2)