cmake: Add cleanup targets clean-cmake-files and clean-all
These are woefully incomplete but they at least get rid of the top level
cmake cruft.  Further work will be needed to locate and remove all the
various generated cmake files.

For now, best practice is to use a parallel build, that can be
recursively deleted when needing to start fresh.

(bzr r14608)