Repair functionality broken by Alex Valvani's "Gtk+ 3 deprecation fixes"
(see in r15547 [1])

Gtk::Label::set_halign() is *not* the same as Gtk::Label::set_xalign() or the deprecated Gtk::Label::set_alignment()
The former sets the alignment of the widget itself while the latter two set the alignment of the text within the widget's size allocation.

Is it really more important to eradicate deprecations then to have working code?

http://bazaar.launchpad.net/~inkscape.dev/inkscape/trunk/revision/15547

(bzr r15580)