cmake/MSYS2: adjust fonts.conf to store fontconfig cache in AppData
MSYS2 uses '/var/cache/fontconfig' by default which is not available outside of the MSYS environment and therefore get's created in the root folder of the Inkscape installation. On most systems this will be in read-only 'Program Files' (which might still work thanks to virtual store, but better be safe then sorry).

This also makes sure we do not re-introduce bug #1196373

(bzr r15661)