SPCanvas::paintSingleBuffer: Draw directly to _backing_store
Before this commit, this function creates a new cairo image surface,
draws the tile that it was told to redraw to this buffer and then uses
cairo to copy the drawn data at the right position inside of
_backing_store. Thus, the drawn data was copied around at least once.

This commit changes the code so that it draws directly to
_backing_store: We get the information from _backing_store and then
create a new image surface from this that covers just the part that we
should redraw. Thus, one copy of the data is avoided.

Signed-off-by: Uli Schlachter <psychon@znc.in>