Move Doxyfile into buildtools/ subdir
doxygen lets you pass the config file name on the command line, so it's
not necessary that this live in the root directory.  We do not host the
doxygen output on the project website, and I'm not aware that many
developers utilize this for their own reference.  However, I've verified
the config file still works, so no reason to get rid of it.