Drop translations of README's
It has been a long time since these were updated, and they're all out of
date anyway due to recent moves to git/cmake/markdown.

If we want to carry translations of high level project information such
as is in the README, there are better places to host it that will be
more translator-friendly.