CI/AppVeyor: Re-introduce full system upgrade
It was disabled in 657285e593dc1b901203bd797c335e87672e3d26
to reduce build times.

Unfortunately pacman does not update dependencies of installed
packages, so if a dependency is already present on the build machine
an outdated (and potentially incompatible) version will be used.

For an example of such breakage, see
https://github.com/Alexpux/MSYS2-packages/pull/1020
https://github.com/appveyor/ci/issues/1822