Unset the current tool early on shutdown.
Before this commit, when the application is terminated we call the
::finish() method for the current tool, only after deleting the
selection and other objects. But it may happen that the tool's finish()
assume that the selection is still alive, making the application crash.
(For instance, with the Bezier tool when the path is still not closed).

Instead, unset the current tool early, when the Desktop object is
removed from the application, before calling its ::destroy() method.