Proper fix for multipage PDF+Latex export
The previous implementation [1] allowed to sandwich text between
graphical objects by outputting multiple (partial) PDF pages which
are then stacked in the final document to reveal the full image.

However this code failed for clipped/masked/transparent objects
as those are treated specially by the renderer resulting in missing
pages in the output causing [2].

The attempt to workaround this issue which was committed in
  e4dea66a338824037b6c35b262aa8db4004b6581 (now reverted)
fixed document creation in LaTeX by inserting blank pages but did
not actually fix the issue with clipped/masked/transparent objects
typically resulting in a single page with the full image and all
text put on top.

This commit resolves the underlying issue, making the former
workaround unnecessary and allowing for proper overlaying of text
and arbitrarily clipped/masked/transparent objects

Fixed bugs:
- https://bugs.launchpad.net/inkscape/+bug/771957 [1]
- https://bugs.launchpad.net/inkscape/+bug/1417470 [2]