Make script warnings transient for "Working..." dialog
This turned out to be a bit of a mess due to the interplay between
the classes Script / ExecutionEnv / Effect / PrefDialog.
(Basically they don't talk to each other much but all want to
 contribute to the GUI)

Likely "Script" (and possibly "ExecutionEnv") should be refactored
to let the other classes handle UI exclusively and throw errors
where suitable.