temporary fix for bug where undo/redo does not restore node-type changes
This is not intended to be a permanent patch, but to point out
the actual source of a problem where undo/redo does not work
for node-type changes.

It works by disabling the function of the PathManipulator::_nodetypesKey()
method, to match some breakage which has apparently occurred somewhere
else. Presumably, somebody would know where that was. Once that is located
and fixed, this patch will be unnecessary, except for a minor code cleanup.
At that point, the functional change would itself constitute a bug, instead
of a temporary fix.

Full details are available at <https://bugs.launchpad.net/inkscape/+bug/1780775>.