Inkscapestream: Fix write methods for string types
As we output UTF8 encoded strings to normal streams we can (and
should) put one byte at a time. Iterating over an ustring will give
us "unichar"s with up to four bytes, though, which is bound to fail
when written to a stream.

Using Glib::ustring::raw() gives us a std::string with the raw
(but still UTF8 encoded) character array that can be iterated
byte-wise and output to the stream as-is.