Purge gunichar from Inkscapestream
None of these functions seems to expect an actual gunichar (which is
a 32-bit type and can hold any UTF-32 or UCS-4 character code, also
known as a Unicode code point).

Instead we want UTF-8 encoded character data (i.e. gchar, which is
equivalent to char) that can be output byte-wise to form a valid
UTF-8 encoded string.