Modify behavior of interpolation
Make sure that the original paths stay the start and end of the
interpolation, even after modifying their shape. This is done by
transforming the trajectory so that it starts at the bounding box
center of the first path and ends at the bounding box center of
the second path.