SPKnot::SPKnot(): improve color style of selected knots
Starting from commit ff04a08000 (working on knots selection, 2017-07-01)
it is possible to select knots for shapes, but the selection style can
be confusing.

The current default colors for knots states are:

  NORMAL    -> white
  MOUSEOVER -> red
  DRAGGING  -> blue
  SELECTED  -> red

Using the same color for the MOUSEOVER and SELECTED states looks like
a contradiction considering that the former is an instantaneous
indicator while the latter represents a more persistent state.

MOUSEOVER and DRAGGING are more similar (in either state the mouse
button is pressed), so switch to the following style:

  NORMAL    -> white
  MOUSEOVER -> red
  DRAGGING  -> red
  SELECTED  -> blue

The new style also matches what path nodes look like.