KnotHolder: fix knots selection status when a knot is clicked without Shift
When there are already selected knots, and  _another_ knot is clicked
without the Shift key pressed, the user would expects the previously
selected knots to be unselected immediately, but this does not happen
until the mouse button is released.

Steps to replicate:
  1. Change size of a rect using one corner.
  2. Change the size using the opposite corner (without pressing Shift).

Observed behavior: The first corner looks still selected during grab.

Expected behavior: The first corner gets unelected as soon as the mouse
                   button is pressed..

This happens because the knots selection status is not updated until the
mouse button is released, i.e. in the click/ungrabbed signal handler.

In order to have a more immediate feedback of the knot selection status,
add a "mousedown" signal handler and update the selection status in
there.

While at it also remove the knot_holder local variable in
knot_clicked_handler() which seems to be unnecessary.