Add a preference to force enable/disable menu icons
By default the theme decides whether it wants to display icons or
not by using the "show-icons" attribute in the menus.xml file.

This attribute is properly inherited an can therefore be used to
selectively enable and disable icons globally (by setting it on the
root element), in certain submenus, or individually per menu item.