Snap: Fix file loading with path with spaces
Currently the inkscape snap is unable to load file with path with
spaces when opening such files by double clicking them in the file
manager(loading them via the file open dialog works).

This is probably due to the fact that the launcher doesn't quote the
expansion of the `@` special parameter and thus all `word`s in the file
path are applied with word separation.  This patch fixes the syntax.

Fixes Launchpad Bug #1735453.

Refer-to: Bug #1735453 “Cannot open files with whitespaces/spaces in pathn...” : Bugs : Inkscape <https://bugs.launchpad.net/inkscape/+bug/1735453>
Refer-to: bash(1) - Parameters - Special Parameters - @
Signed-off-by: 林博仁(Buo-ren, Lin) <Buo.Ren.Lin@gmail.com>