Fix query for monitor dimensions
The call to
  Gdk::Display::get_monitor_at_window()
in
  Inkscape::UI::get_monitor_geometry_at_window()
requires the underlying GdkWindow to be fully initialized.

This is achieved by calling "realize()" before attempting to read
that information.

Previously we used to show the window first (which implies realizing
it) which worked around the issue. However it required us to hide
and show it once again later if we wanted to move it in order for
the window manager's routines to sanitize the position to kick in.