Save default template as "default.$lang.svg" if present
By default we want to save as "default.svg", so it works indepentent
from the UI language selected.

However if a localized default template like "default.de.svg" is
already present (i.e. manually created by the user), it will
override our custom template, so we need to overwrite it.

Second  part of a fix for
  https://gitlab.com/inkscape/inkscape/issues/103