Remove custom tab order in Document Properties
The method for handling custom tab order in GTK (set_focus_chain) has
been deprecated since GTK 3.24 with no direct replacement.

I've decided to remove this code as the code did not seem to make any
difference during testing: the tab order appeared to be the same.

Regardless, 'Page -> Custom size' in Document Properties is the only
place that it's used, so the impact should be fairly limited anyway.