Remove _argv0 from Inkscape::Application
Currently it was only used as the least preferred (and therefore
effectively unused) location where to store the crash backup.

However it wasn't set properly since
  408cb49b5559a81ea803df64bf58457a5dd4bf16
causing assertion errors while crashing.

On top of that argv0 is not a reliable way to determine the path to
the currently running executable anyway.

Fixes https://gitlab.com/inkscape/inkscape/issues/176