Preferences: add methods to set/get unsigned integers
Those should be used for RGBA color values, which would overflow
the signed type otherwise.

Replacement will be seamless and fully backwards-compatible, as we
can read the overflowed (i.e. negative) signed integers we used to
save and they will be properly unwrapped.

Also reading the unsigned ints as signed ints in older versions of
Inkscape will still work, as we always depended on the undefined
behavior of atoi in case of out of range numbers (which luckily is
means overflowing in gcc).