Add appearance="colorbutton" to parameters of type color
Renders a simple Gtk::ColorButton with label instead of the full
ColorNotebook which is too large for many applications.

Should likely be re-implemented bringing up a new Dialog that wraps
an Inkscape::UI::Widget::ColorNotebook eventually, instead of
relying on the Gtk::ColorButton's picker implementation which is
inconsistent to existing pickers.