Change build OS and target
We no longer user El Capitan (10.11) as build platform. We now use
Mojave (10.14) in combination with the Mavericks SDK (10.9). This
simplifies the build process as we can use a current platform
and gives backward compatibility down to Mavericks (10.9).

Use a newer Python build that we turn non-relocatable for JHBuild,
avoiding to build it ourselves (time consuming!) in 120-python3.sh.

New default build location is /Users/Shared/work. This is a user-
independent location and user-writable and simplifies WKR_DIR
handling.