Extensions: Implement translationdomain functionality
Inkscape will read the "translationdomain" attribute from the
<inkscape-extension> root element in the .inx file.

It will then attempt to lookup a message catalog that matches
this domain, register it with gettext, and use it for translations.

Message catalogs may be located in either
- the .inx file's location
- the root of the "extensions" directory containing the .inx
- the system location Inkscape's own catalog is loaded from

To make this functionality available to script extensions, Inkscape
will set the environment variables
  INKEX_GETTEXT_DOMAIN
and
  INKEX_GETTEXT_DIRECTORY
so extension scripts can then use something like
  bindtextdomain(INKEX_GETTEXT_DOMAIN, INKEX_GETTEXT_DIRECTORY)
  textdomain(INKEX_GETTEXT_DOMAIN)
to enable the feature.

See also
  https://gitlab.com/inkscape/inkscape/issues/333
  https://gitlab.com/inkscape/extensions/issues/117