Fixed pen tilt computation in Calligraphy tool
This fixes incorrect pen angle in Calliraphy tool with "Use the tilt of
the input device to alter the angle of the pen's nib" enabled.

Before the tool's angle would rotate in opposite direction than
the stylus.

I did test with the gtk3-demo app that the reported input device
tilt is correct.