Prevent line height change on tool switch
Changing line height: When switching to another tool with the selection still active, line height is applied to the entire text
When the user switch to another tool from text tool text selection is fired and get wrong in some circunstancies with line height.
To fix it we check in the function that change it if the current tool is text tool, es we do in similar function in the file.

Fix issue https://gitlab.com/inkscape/inkscape/-/issues/1354