Fix Stroke to Path for objects with transform
Before, if the object had an explicit "transform" attribute, this
transform was effectively applied twice: Once, baked into the path
generated for the marker; and again as an explicit transform applied to
the newly created group that encloses everything.

This change removes the transform baked into the marker's generated
path. This matches how the transform also isn't baked into the stroke's
generated path.

Objects with explicit "transform" attributes can happen either due to
the "Store transformation: Preserved" preference or when an object just
can't have transforms baked in (like an arc).

Resolves: #2709