Fix rendering issue with a powerstroke join
In 1.0 the extrapolated arc join of the powerstroke LPE renders
incorrectly - there is a "pinprick" on the outer edge and a
discontinuity that is sometimes obvious. This did not happen in 0.92.5,
and is fixed here.

This fixes https://gitlab.com/inkscape/inbox/-/issues/3289. The issue
was discovered when doing https://derpibooru.org/images/2400131.