Auto-size "Move to Layer" dialog to its contents
When layers with long names are used, the hard-coded minimum size
of the "Move to Layer" dialog of 220x180 pixels often requires
resizing the (initially tiny) dialog or scrolling its contents
horizontally. Both are awkward and unnecessary, especially since
the dialog is modal anyway.

This change tells the scroller to propagate the layer tree's
initial natural dimensions to the dialog, making it automatically
sized so that all of the initially expanded layers are visible.

Resolves: inkscape/inbox#2541