Fix node tool crash with LPE
Fix a crash when Node tool is changed to from other tools and it has a LPE item selected
It happends on new documents because in master LPE calculation is made only on user iteraction instead do on load file. this allow faster open in docs with lots of LPE items.
This MR is a fix continuation of this commit with a derived bug.

https://gitlab.com/inkscape/inkscape/-/commit/c104d994561d73c7a2399f3040607a8bdd488161