Ungrouping: Fix transform of linked non-children
Transform of linked items must only be compensated if their source is
a direct child of the group being ungrouped. This also applies to
linked items which are not ungrouped themselves but their source is.

This was already working for clones, so the fix is to handle all
clone-like links the same.