Fix build with glib 2.67.3.
As of glib 2.67.3, <glib.h> can no longer be included in extern "C"
blocks. It was indirectly included by both "types.h" and "color.h".
"types.h" already does not need to be wrapped in an extern "C" block,
"color.h" does but can be modified not to, so with that changed they can
be moved out.