Improve theming to allow user icon and theme (in gtk) bypass our theming system
To do this we check the preferences of both Gtk settings are not set, in this case internaly
we use the default icon/theme set on startup when necessary but don`t call to a theming gtk settings update.
We also store all necessary default preferences to allow restore user launch settings
By the way I fix welcome screen setting icon theme to Adwaita.
Adwaita is the default not fallback theme on GTK and always is present in GTK3 package, if we use for icons and the system has other icon theme active there is not problems because both themes fallback to hicolor.
Finaly using hicolor as theme name produce wired things so I ensure hicolor is never set (by Adwaita substitution).